// WYNKWeb/src/app/Models/StockmasterModel.ts
export class StockmasterModel {
  GRN: string;
  SMID: number;
  TransactionID: number;
  CMPID: number;
  DocumentNumber: string;
  DocumentDate: Date;
  QtnNumber: string;
  QtnDate: Date;
  POID: number;
  StoreID: number;
  TransactionType: string;
  VendorID: number;
  GrossProductValue: number;
  TotalDiscountValue: number;
  TotalTaxValue: number;
  TotalCGSTTaxValue: number;
  TotalSGSTTaxValue: number;
  TotalPOValue: number;
  IsCancelled: boolean;
  TermsConditions: number;
  IsDeleted: boolean;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  DCNumber: string;
  DCDate: Date;
  TotalIGSTTaxValue: number;
  CreatedUTC: Date;

}