// WYNKWeb/src/app/Models/ItemBatchModel.ts
export class ItemBatchModel {

  ItemBatchID: number;
  ItemID: number;
  ItemBatchNumber: string
  ItemBatchQty: number;
  ItemBatchissueQty: number;
  ItemBatchBalnceQty: number;
  ItemBatchExpiry: Date
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  CreatedUTC: Date;
  StoreID: number;
}