// WYNKWeb/src/app/Models/ItemBatchTranModel.ts
export class ItemBatchTranModel {


  ItemBatchTransID: number;
  ItemBatchID: number;
  TC: number;
  SMID: number
  STID: number;
  ItemID: number;
  ItemBatchNumber: string;
  ItemBatchTransactedQty: number;
  ItemBatchExpiry: Date
  UOMID: number;
  CreatedUTC: Date;
  ContraItemBatchID: number;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;

}